+++ src/main/scala/neighbourhood/FirstImprovementSelector.scala
@@ -6,11 +6,13 @@ package neighbourhood
   * @constructor create a new first improvement selector.
   * @param costFunction cost function to maximize
   */
-class FirstImprovementSelector[T](override val costFunction: List[T] => Int) extends Selector(costFunction) {
+class FirstImprovementSelector[T](
+  override val costFunction: List[T] => Double
+) extends Selector(costFunction) {
 
   /** @see neighbourhood.Selector.select() */
   def select(refElem: List[T], neighbourhood: Stream[List[T]]): List[T] = {
-    def innerSelect(refScore: Int, refE: List[T], n: Stream[List[T]]): List[T] = {
+    def innerSelect(refScore: Double, refE: List[T], n: Stream[List[T]]): List[T] = {
       if (n.isEmpty)
         refE
       else {
